Test code now demonstrates label
