+++ base/uk.ac.stfc.isis.ibex.scriptgenerator.client/Application.e4xmi
@@ -4,7 +4,9 @@
     <tags>shellMaximized</tags>
     <children xsi:type="advanced:PerspectiveStack" xmi:id="_uMW6EJGoEemDrc7_A6qM8g" elementId="uk.ac.stfc.isis.ibex.scriptgenerator.client.perspectivestack.0">
       <children xsi:type="advanced:Perspective" xmi:id="_umobQJGoEemDrc7_A6qM8g" elementId="uk.ac.stfc.isis.ibex.scriptgenerator.client.perspective.0">
-        <children xsi:type="basic:Part" xmi:id="_8NgLsJHLEemDrc7_A6qM8g" elementId="uk.ac.stfc.isis.ibex.scriptgenerator.client.part.scriptingview" contributionURI="bundleclass://uk.ac.stfc.isis.ibex.ui.scriptgenerator/uk.ac.stfc.isis.ibex.ui.scriptgenerator.views.TableView" label="scripting view"/>
+        <children xsi:type="basic:PartStack" xmi:id="_Co6VgJNkEemAXoaj-ZzBLA" elementId="uk.ac.stfc.isis.ibex.scriptgenerator.client.partstack.0">
+          <children xsi:type="basic:Part" xmi:id="_8NgLsJHLEemDrc7_A6qM8g" elementId="uk.ac.stfc.isis.ibex.ui.scriptgenerator.views.TestView" contributionURI="bundleclass://uk.ac.stfc.isis.ibex.ui.scriptgenerator/uk.ac.stfc.isis.ibex.ui.scriptgenerator.views.TestView" label="scripting view"/>
+        </children>
       </children>
     </children>
     <mainMenu xmi:id="_zGFwsJHKEemDrc7_A6qM8g" elementId="uk.ac.stfc.isis.ibex.scriptgenerator.client.menu.0">
