+++ base/uk.ac.stfc.isis.ibex.scriptgenerator.client/META-INF/MANIFEST.MF
@@ -2,7 +2,8 @@ Manifest-Version: 1.0
 Bundle-SymbolicName: uk.ac.stfc.isis.ibex.scriptgenerator.client;singleton:=true
 Bundle-Name: IBEX
 Bundle-Version: 1.0.0.qualifier
-Require-Bundle: uk.ac.stfc.isis.ibex.scriptgenerator.product;bundle-version="1.0.0"
+Require-Bundle: uk.ac.stfc.isis.ibex.scriptgenerator.product;bundle-version="1.0.0",
+ uk.ac.stfc.isis.ibex.ui.scriptgenerator;bundle-version=1.0.0
 Bundle-ManifestVersion: 2
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Vendor: ISIS
