+++ base/uk.ac.stfc.isis.ibex.ui.scriptgenerator/META-INF/MANIFEST.MF
@@ -9,8 +9,6 @@ Require-Bundle: uk.ac.stfc.isis.ibex.scriptgenerator;bundle-version="1.0.0",
  org.eclipse.core.runtime;bundle-version="3.8.0",
  uk.ac.stfc.isis.ibex.model;bundle-version="1.0.0",
  org.eclipse.swt,
- uk.ac.stfc.isis.ibex.validators;bundle-version="1.0.0",
- uk.ac.stfc.isis.ibex.ui.widgets;bundle-version="1.0.0",
  org.eclipse.jface,
  org.eclipse.ui.workbench,
  uk.ac.stfc.isis.ibex.ui;bundle-version="1.0.0"
