--- base/uk.ac.stfc.isis.ibex.ui.scriptgenerator/src/uk/ac/stfc/isis/ibex/ui/scriptgenerator/views/TableView.java
@@ -1,74 +1,0 @@
-
-/*
-* This file is part of the ISIS IBEX application.
-* Copyright (C) 2012-2015 Science & Technology Facilities Council.
-* All rights reserved.
-*
-* This program is distributed in the hope that it will be useful.
-* This program and the accompanying materials are made available under the
-* terms of the Eclipse Public License v1.0 which accompanies this distribution.
-* EXCEPT AS EXPRESSLY SET FORTH IN THE ECLIPSE PUBLIC LICENSE V1.0, THE PROGRAM 
-* AND ACCOMPANYING MATERIALS ARE PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES 
-* OR CONDITIONS OF ANY KIND.  See the Eclipse Public License v1.0 for more details.
-*
-* You should have received a copy of the Eclipse Public License v1.0
-* along with this program; if not, you can obtain a copy from
-* https://www.eclipse.org/org/documents/epl-v10.php or 
-* http://opensource.org/licenses/eclipse-1.0.php
-*/
-
-package uk.ac.stfc.isis.ibex.ui.scriptgenerator.views;
-
-import java.util.ArrayList;
-
-import javax.annotation.PostConstruct;
-
-import org.eclipse.swt.SWT;
-import org.eclipse.swt.layout.GridData;
-import org.eclipse.swt.layout.GridLayout;
-import org.eclipse.swt.widgets.Composite;
-
-import uk.ac.stfc.isis.ibex.scriptgenerator.PythonBuilder;
-import uk.ac.stfc.isis.ibex.scriptgenerator.row.Row;
-import uk.ac.stfc.isis.ibex.scriptgenerator.settings.ApertureSans;
-import uk.ac.stfc.isis.ibex.scriptgenerator.settings.ApertureTrans;
-import uk.ac.stfc.isis.ibex.scriptgenerator.settings.CollectionMode;
-import uk.ac.stfc.isis.ibex.scriptgenerator.settings.Order;
-import uk.ac.stfc.isis.ibex.scriptgenerator.settings.SampleGeometry;
-import uk.ac.stfc.isis.ibex.scriptgenerator.settings.SansSettings;
-import uk.ac.stfc.isis.ibex.ui.scriptgenerator.Table;
-
-/**
- * A panel to hold a ScriptGeneratorTable.
- */
-public class TableView {
-    /** The class ID string. **/
-	public static final String ID = "uk.ac.stfc.isis.ibex.ui.scriptgenerator.scriptgeneratorview";
-    /** The table. */
-    private Table table;
-    
-	@PostConstruct
-	@SuppressWarnings("checkstyle:magicnumber")
-	public void createPartControl(Composite parent) {
-		parent.setLayout(new GridLayout(2, true));
-		SansSettings settings = new SansSettings(1, 1, 7, 7, Order.TRANS, false, ApertureSans.MEDIUM, ApertureTrans.MEDIUM, SampleGeometry.DISC, CollectionMode.HISTOGRAM);
-		
-		ArrayList<Row> rows = new ArrayList<Row>();
-        rows.add(new Row());
-        
-		PythonBuilder builder = new PythonBuilder();
-		builder.setSettings(settings);
-        builder.setRows(rows);
-        
-        table = new Table(parent, SWT.NONE, SWT.MULTI | SWT.NO_SCROLL | SWT.V_SCROLL | SWT.FULL_SELECTION | SWT.BORDER);
-		table.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 2, 1));
-		table.setRows(rows);
-	}
-	
-	/**
-	 * Clears the table.
-	 */
-	public void clearTable() {
-		table.clearTable();
-	}
-}
